+++ jest_env_setup.js
@@ -13,3 +13,4 @@ import translations from "translations.json";
 window.I18n = new I18n(translations);
 
 global.DOCS_URL = "http://localhost:4000";
+global.RELATIVE_URL_ROOT = "";